requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.7.14
python-dotenv==1.1.1
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
